// src/models/List.model.ts
export interface ListModel {
  id: number
  rowName: string
  total?: number
  salary: number
  mimExploitation: number
  machineOperatorSalary: number
  materials: number
  mainCosts: number
  supportCosts: number
  equipmentCosts: number
  overheads: number
  estimatedProfit: number
  child: ListModel[]
}